Use standard amqp and management ports in docker tests.
We originally forwarded the amqp and management from the docker container
to the host on the nonstandard ports 5673 and 15673, respectively.  This
was done in case there was a RabbitMQ instance already running on the host.
However, that is probably unlikely, and it causes confusion in the test
environment.

If we do want to use nonstandard ports in the future, we'll have to ensure
we have the test environment set up properly.