Bug 1335954: Add RESERVED_USERS_{REGEX,MESSAGE}
The current reason for these options is to prohibit management of users
beginning with "taskcluster-", since those will be handled by tc-pulse
(which has a similar configuration to prevent it handling anything *not*
beginning with that prefix).

The design is sufficiently general that it could be used to carve out
other parts of the user namespace for other purposes, if that became
necessary.