Bug 1378380 - Add CSRF protection to DELETE calls.
We have to use a slightly different method here than is used in the POST
calls, since there is no form element.  Instead, we send the token back
to the server as a header.

Arguably, for simplicity, headers should be used in all cases, but that
isn't important to fix now.