Bug 1362106 - Fix queues_listing call for admins.
When logged in as a admin, the queues_listing view was only providing the
current user and unowned queues.

Also changed the template to not depend specifically on whether
or not the user is an admin, but rather on how many users are
passed in.  This is the important point; the template doesn't
need to care whether the user is an admin for this.