Remove ssl-pair-generation monkey patch.
We worked around an old problem in werkzeug in which the key
length for the adhoc ssl certificates, used in testing, were
too short after a Firefox update.  This was rectified some
time ago and hence our monkey patch is no longer necessary
(and in fact now causes other problems).