All admins to delete Pulse users from all_pulse_users.html.
This works pretty much the same way as deleting users on
profile.html.  However, the slideUp() command, used to remove
the user listing without having to refresh, doesn't reset the
table row count at the bottom, which isn't ideal.  Better would
be to use the DataTables API to remove the row.  Even better
would be to unify the two views to display in roughly the same
way.

For now, this is fine, as this function is only available to
admins anyway.